+++ Okta/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2025-09-16 - 2.8.1
+
+### Changed
+
+- Add user and device asset connectors
+
 ## 2024-05-28 - 2.8.0
 
 ### Changed