+++ Tenable/CHANGELOG.md
@@ -7,12 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## 2025-08-15 - 1.0.1
+## 2025-09-30 - 1.0.2
 
 ### Changed
 
 - Update the sdk version to 1.21.0
 - Add batch size configuration to the connector settings
+
+## 2025-09-19 - 1.0.1
+
+### Changed
+
+- Update asset connector name
+- Add module run method
+
 
 ## 2025-08-15 - 1.0.0